Input name bugfix in runner_utils
There are cases where the names of the inputs
in the edge-IR graph do not match the names of the inputs
in the export-IR. Since desc.json used the names from export-IR
while TOSA used the names from edge-IR there was a mismatch.

I therefore  changed desc.json to use the names from edge-IR if possible.
I also changed the search of tosa.fbs to be cwd independent,
and added a check to avoid a crash if you .dump_artifact() on a
graph with no delegate.

Signed-off-by: Erik Lundell <erik.lundell@arm.com>
Change-Id: Iffa56e2f43910adc74608b96518717d14e0beb53