[executorch] Migrate runtime/core to new namespace
Pull Request resolved: https://github.com/pytorch/executorch/pull/4607

Migrate these headers to the new `::executorch::runtime` namespace. Add temporary aliases from the old `::torch::executor` namespace so we can migrate users incrementally.

The tests for these headers are updated in the next diff to help demonstrate that the backwards-compatibility aliases work.

Because `EValue` changed namespaces, I also needed to update print_evalue: the `operator<<()` for a type needs to live in its primary namespace. This shows that the aliases aren't perfect, but very few other users should be defining their own `operator<<` for our types. This is the only scenario where I've seen the namespace aliases break down.
ghstack-source-id: 238157921
@exported-using-ghexport

Differential Revision: [D60428951](https://our.internmc.facebook.com/intern/diff/D60428951/)