[XNNPACK][Partitioner] enable src based partitioner
Pull Request resolved: https://github.com/pytorch/executorch/pull/4762

In order to maintain parity with the current to_edge and to_backend lowering flow, we need to support source based partitioning.

We apply source-based partitioning to the AddMMConfig to partition all the nodes surrounding addmm so that it can be recomposed internally. While this is fine for to_edge and to_backend flow. For more robust flow, we will not have to use this when running to_edge_transform_and_lower.
ghstack-source-id: 238765476
@exported-using-ghexport

Differential Revision: [D61250576](https://our.internmc.facebook.com/intern/diff/D61250576/)