[ET-VK][ez] Empty initialize ShaderInfo and add `bool()` operator
## Context

This diff lays the foundation for the implementation of "view operators". These operators will not dispatch any shaders; they will be solely responsible for updating the sizes and strides metadata of the output tensor, which will use the same storage resource as the input tensor. These ops will be implemented by adding a "no-op" `ExecuteNode` instance to the `ComputeGraph`, which does not contain a shader but does contain a resize function to update sizes and strides upon a resize.

This diff allows `ShaderInfo` to empty initialize, and add a bool operator to check if the ShaderInfo actually points to valid shader code. This will be used to construct "no-op" `ExecuteNode` instances which can be used to implement view operators.

Differential Revision: [D61666460](https://our.internmc.facebook.com/intern/diff/D61666460/)

ghstack-source-id: 239353688
Pull Request resolved: https://github.com/pytorch/executorch/pull/4842