[ET-VK] Add `TmpTensor` struct to recycle temporary tensor memory
Pull Request resolved: https://github.com/pytorch/executorch/pull/5041

## Context

Normally, tensor memory is planned during the export stage; tensors that do not overlap in lifetimes may share a memory allocation. However, memory planning requires knowledge of the lifetime of the tensors.

However, some complex operators may not be able to perform all the necessary computations in one shader, or the implementation of the operator may require that some temporary tensors be created during the execution of the op. Since these temporary tensors are not visible to the memory planning algorithm, they will not be memory planned.

This diff introduces the `TmpTensorVRef` object which facilitates memory sharing between temporary tensors. The design principle is that the lifetime of temporary tensors is restricted to the execution of the op within which they are created; thus, that knowledge can be used to implement memory planning. Please see the comments documentation of `TmpTensorVRef` for more details.
ghstack-source-id: 241061239
@exported-using-ghexport

Differential Revision: [D62144398](https://our.internmc.facebook.com/intern/diff/D62144398/)