[ET-VK] Introduce axis mapping for no-copy permute of texture-backed tensors
Pull Request resolved: https://github.com/pytorch/executorch/pull/5092

## Context

This diff introduces the `axis_mapping` field for `vTensors`, which can be used to implement no-copy permutes. The idea behind the axis mapping is that it is somewhat analogous to dim order for texture backed tensors.

The axis mapping is normalized to 4 dimensions, similar to padded sizes. The first 3 elements indicates which of the (X,Y,Z) image texture axes the width, height, and channels dim of the tensor maps to. The final element indicates the WHCN index of the tensor dimension along which batches will be concatenated.

The benefit of introducing axis mapping is twofold:

1. Permutes can be performed without any data copying by re-using a texture but updating the axis mapping.
2. Allows the memory layout of texture backed tensors to be more flexible, and optimize for performance or memory footprint by using unconventional axis mappings.

Regarding the second point, we have found that adding length to a texture's Z axis is more costly than adding length to the texture's X or Y axes. Similarly, we have found that reading along the Z axis yeilds slightly lower throughput than reading along the X or Y axes. By introducing axis mapping, we can map the largest dimension to a texture's X axis instead of mapping it to the most intuitive texture axis (i.e. channels to Z axis). This can save a lot of texture memory and potentially improve compute shader latency as well.

However, the pre-requisite of using texture mapping heavily is that the overhead introduced in calculating tensor indices and texture positions does not significantly increase compute shader latency. The impact of this will be investigated and shown in the following diffs.

Note that this diff only introduces the `axis_mapping` field;
ghstack-source-id: 241282077
@exported-using-ghexport

Differential Revision: [D62210118](https://our.internmc.facebook.com/intern/diff/D62210118/)